Fix fab manager config compatibility

+++ backend/src/main/java/com/tsmc/lims/backend/controller/NotificationController.java
@@ -32,7 +32,7 @@ public class NotificationController {
     }
 
     @DeleteMapping("/{notificationId}")
-    public void deleteNotification(@RequestParam String employeeId, @PathVariable Long notificationId) {
+    public void deleteNotification(@RequestParam String employeeId, @PathVariable Integer notificationId) {
         service.deleteNotification(employeeId, notificationId);
     }
