+++ backend/src/main/java/com/tsmc/lims/backend/repository/NotificationRepository.java
@@ -4,10 +4,10 @@ import com.tsmc.lims.backend.domain.NotificationEntity;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<NotificationEntity, Long> {
+public interface NotificationRepository extends JpaRepository<NotificationEntity, Integer> {
     List<NotificationEntity> findByUserEmployeeIdOrderByCreatedAtDesc(String employeeId);
 
-    void deleteByNotifIdAndUserEmployeeId(Long notifId, String employeeId);
+    void deleteByNotifIdAndUserEmployeeId(Integer notifId, String employeeId);
 
     void deleteByUserEmployeeId(String employeeId);
 }
