+++ backend/src/main/java/com/tsmc/lims/backend/repository/WaferRepository.java
@@ -4,6 +4,6 @@ import com.tsmc.lims.backend.domain.WaferEntity;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WaferRepository extends JpaRepository<WaferEntity, Long> {
+public interface WaferRepository extends JpaRepository<WaferEntity, Integer> {
     List<WaferEntity> findByRequestRequestIdOrderByWaferCode(String requestId);
 }
