+++ backend/src/main/java/com/tsmc/lims/backend/repository/WipTaskRepository.java
@@ -4,7 +4,7 @@ import com.tsmc.lims.backend.domain.WipTaskEntity;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WipTaskRepository extends JpaRepository<WipTaskEntity, Long> {
+public interface WipTaskRepository extends JpaRepository<WipTaskEntity, Integer> {
     boolean existsByRequestRequestId(String requestId);
 
     List<WipTaskEntity> findByStatusOrderByDispatchedAtAsc(String status);
